[AV1][Coverity] Fix INTEGER_OVERFLOW CID 4398015 in order_hint_bits_minus1
GetBits(3) returns uint32_t. In the ternary expression, C++ type
promotion makes the result type uint32_t, converting -1 to 4294967295U
before assignment to int32_t order_hint_bits_minus1. Use
static_cast<int32_t> to preserve signed semantics.

Coverity CID: 4398015

+++ _studio/shared/umc/codec/av1_dec/src/umc_av1_bitstream.cpp
@@ -1421,7 +1421,7 @@ namespace UMC_AV1_DECODER
                 sh.seq_force_integer_mv = 2;
 
             sh.order_hint_bits_minus1 =
-                sh.enable_order_hint ? GetBits(3) : -1;
+                sh.enable_order_hint ? static_cast<int32_t>(GetBits(3)) : -1;
         }
 
         sh.enable_superres = GetBit();
